See electrical meeting agendas 3/27/26
Created new branch away from merge issues and verified that temperature feedback works on ESP monitor (but not UART python script terminal).

+++ ESP-Projects/main/user_libs/inc/packets.h
@@ -43,12 +43,12 @@ typedef struct __attribute__((packed))
     uint8_t reserved_bit1;
     uint8_t reserved_bit2;
     
-    float front_left_wheel_temp;
-    float back_left_wheel_temp;
-    float front_right_wheel_temp;
-    float back_right_wheel_temp;
-    float front_drum_temp;
-    float back_drum_temp;
+    int front_left_wheel_temp;
+    int back_left_wheel_temp;
+    int front_right_wheel_temp;
+    int back_right_wheel_temp;
+    int front_drum_temp;
+    int back_drum_temp;
 } TempPacket_OneRobot;
 
 typedef struct __attribute__((packed))
